Update categories API
